refactor(internal): keep loader-side atoms out of component files
- The router keeps every page's loader in the main bundle, so what a loader imports reaches every visitor, sign-in page included. The research inbox, the run list and the company panels each exported the atom or dialog schema their loader needed, which dragged those whole screens and their stylesheets onto the sign-in page.
- Those pieces now live in `atoms/research-inbox-atoms.ts`, `atoms/research-atoms.ts`, `lib/research-dlg.ts` and `lib/company-dlg.ts`; measured on the sign-in page, scripts went from 738 kB to 408 kB and stylesheets from 18 to 6.
- The runs loader now hands its data over outside its catch, like the other loaders, so a missing serialization key breaks the page instead of quietly turning into a refetch.
